requirements.txt
Pillow
heic2png
